requirements.txt
nltk==3.5
pandas==1.2.3
grpcio==1.36.1
numpy==1.20.1
protobuf==3.15.8
python-dotenv==0.17.0
